--- a/pkg/convert/parser_test.go
+++ b/pkg/convert/parser_test.go
@@ -26,7 +26,7 @@ var _ = Describe("convert", func() {
 		It("parses data correctly", func() {
 			result := []string{}
 
-			b, err := ioutil.ReadFile("fixtures/cpu.pprof")
+			b, err := ioutil.ReadFile("testdata/cpu.pprof")
 			Expect(err).ToNot(HaveOccurred())
 			r := bytes.NewReader(b)
 			g, err := gzip.NewReader(r)