… go.sum
… pkg/convert/profile-extensions.go
… pkg/convert/testdata/cpu.pprof