The PR is aimed to improve pyroscope server and profiler performance. Primarily by reducing the amount and size of allocations.

1. Use byte buffer pool for serialization.

Although it might be not so apparent in synthetic benchmarks, this should alleviate GC work. The benchmark shows that with a buffer pool, we decrease size of allocations by more than 20% (example for an empty `Segment`):
```
BenchmarkSegment_Serialize/Before-4            370618              2828 ns/op            1288 B/op         22 allocs/op
BenchmarkSegment_Serialize/After-4             397603              2807 ns/op            1032 B/op         19 allocs/op
```

2. Refactor `GoSpy` pprof profile traversal.

This change affects both server and agent ends. Before the change, the function could constitute up to 20-50% of all allocations (by size).
```
BenchmarkProfile_Get/Before-4              77934             14211 ns/op            8094 B/op        105 allocs/op
BenchmarkProfile_Get/After-4              350104              3637 ns/op             429 B/op          3 allocs/op
```

Related: https://github.com/pyroscope-io/pyroscope/issues/294