@olvrng 2021-08-02
I tested `sync.Pool` and `bytebufferpool` by running benchmark from this repository: [omgnull/go-benchmark](https://github.com/omgnull/go-benchmark) (the benchmark is recommended by [bytebufferpool](https://github.com/valyala/bytebufferpool)).

Here is the result on my machine with Go 1.14 (the version from our [go.mod](https://github.com/pyroscope-io/pyroscope/blob/main/go.mod))

```
$ go test ./... -bench=. -benchmem
?   	github.com/omgnull/go-benchmark/buffer	[no test files]
goos: darwin
goarch: amd64
pkg: github.com/omgnull/go-benchmark/buffer/main
cpu: Intel(R) Core(TM) i7-6567U CPU @ 3.30GHz
BenchmarkGenericBuf-4                       	 2474088	       486.2 ns/op	    2576 B/op	       4 allocs/op
BenchmarkGenericStackBuf-4                  	 2242417	       644.1 ns/op	    2576 B/op	       4 allocs/op
BenchmarkAllocBuf-4                         	 2161522	       488.1 ns/op	    2576 B/op	       4 allocs/op
BenchmarkSyncPoolBuf-4                      	21519672	        49.74 ns/op	       0 B/op	       0 allocs/op
BenchmarkBpoolPoolBuf-4                     	 5466226	       218.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkByteBufferPoolBuf-4                	19147827	        54.19 ns/op	       0 B/op	       0 allocs/op
BenchmarkEasyJsonBuffer-4                   	 4705569	       247.9 ns/op	     600 B/op	       4 allocs/op
BenchmarkEasyJsonBuffer_OptimizedConfig-4   	11684805	        97.13 ns/op	      24 B/op	       1 allocs/op
PASS
ok  	github.com/omgnull/go-benchmark/buffer/main	12.111s
```

It seems `sync.Pool` is faster than `bytebufferpool`. So I suggest removing the package and just use `sync.Pool` from the standard library.