@kolesnikovae 2021-08-02
To me it looks more like fluctuations - see my results below:

`go 1.14.15`:
```
goos: darwin
goarch: amd64
BenchmarkGenericBuf-4                            1550229               765 ns/op            2576 B/op          4 allocs/op
BenchmarkGenericStackBuf-4                       1504646               793 ns/op            2576 B/op          4 allocs/op
BenchmarkAllocBuf-4                              1524476               815 ns/op            2576 B/op          4 allocs/op
BenchmarkSyncPoolBuf-4                          16666263                81.6 ns/op             0 B/op          0 allocs/op
BenchmarkBpoolPoolBuf-4                          4400764               269 ns/op               0 B/op          0 allocs/op
BenchmarkByteBufferPoolBuf-4                    16411653                75.8 ns/op             0 B/op          0 allocs/op
BenchmarkEasyJsonBuffer-4                        3181826               371 ns/op             608 B/op          4 allocs/op
BenchmarkEasyJsonBuffer_OptimizedConfig-4        7567078               139 ns/op              32 B/op          1 allocs/op
PASS
ok      go-benchmark/buffer/main     13.804s
```

`go 1.16.3`:
```
goos: darwin
goarch: amd64
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
BenchmarkGenericBuf-4                            1578630               738.7 ns/op          2576 B/op          4 allocs/op
BenchmarkGenericStackBuf-4                       1441327               695.4 ns/op          2576 B/op          4 allocs/op
BenchmarkAllocBuf-4                              1619761               755.8 ns/op          2576 B/op          4 allocs/op
BenchmarkSyncPoolBuf-4                          17595960                74.40 ns/op            0 B/op          0 allocs/op
BenchmarkBpoolPoolBuf-4                          4571432               264.8 ns/op             0 B/op          0 allocs/op
BenchmarkByteBufferPoolBuf-4                    18865262                58.98 ns/op            0 B/op          0 allocs/op
BenchmarkEasyJsonBuffer-4                        3780392               309.9 ns/op           600 B/op          4 allocs/op
BenchmarkEasyJsonBuffer_OptimizedConfig-4        9238102               121.0 ns/op            24 B/op          1 allocs/op
PASS
ok      go-benchmark/buffer/main     12.559s
```

`go1.17-b8ca6e5` (tip):
```
goos: darwin
goarch: amd64
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
BenchmarkGenericBuf-4                            1604217               799.3 ns/op          2576 B/op          4 allocs/op
BenchmarkGenericStackBuf-4                       1449591               813.2 ns/op          2576 B/op          4 allocs/op
BenchmarkAllocBuf-4                              1501189               796.2 ns/op          2576 B/op          4 allocs/op
BenchmarkSyncPoolBuf-4                          16659342                61.95 ns/op            0 B/op          0 allocs/op
BenchmarkBpoolPoolBuf-4                          4727404               246.3 ns/op             0 B/op          0 allocs/op
BenchmarkByteBufferPoolBuf-4                    17938633                58.09 ns/op            0 B/op          0 allocs/op
BenchmarkEasyJsonBuffer-4                        3227756               344.5 ns/op           600 B/op          4 allocs/op
BenchmarkEasyJsonBuffer_OptimizedConfig-4        8777935               128.8 ns/op            24 B/op          1 allocs/op
PASS
ok      go-benchmark/buffer/main     12.178s
```

But this benchmark suite uses a static data set which does not really allow evaluating `bytebufferpool` advantages. One great think about `bytebufferpool` (and the reason to replace `sync.Pool`) is that due to its internal calibrating mechanism it is more adaptive, which results in limited memory fragmentation and more efficient allocations. 

BTW, re go version - I think it's time to switch to newer one.