@petethepig 2021-08-03
@kolesnikovae re go version — I updated it for linux and windows https://github.com/pyroscope-io/pyroscope/commit/e44f8c8d23467528a7ccb9846fee9d033529020f https://github.com/pyroscope-io/pyroscope/commit/94d5c4323dc8c3f18214bbde8ff1a903e551b993